Use autofs instead of entries in /etc/fstab
This allows for remote filesystems to be lazily mounted and for all nodes in
the cluster to be started simultaneously.

The "mh-opsworks-recipes::nfs-client" recipe should be run during the
configure part of the opsworks lifecycle.

Includes:

* Updates /etc/auto.master
* Creates /etc/matterhorn, which includes the entry to mount remote storage
* Reloads the autofs service
* Provision storage from external sources as well as the Storage layer

See the "NFS Storage Options" section of the mh-opsworks README for more
information about how to use external NFS storage.
https://github.com/harvard-dce/mh-opsworks