Fully enable enhanced networking on ubuntu 14.04
This will compile and install the driver necessary to get full 10Gbps
networking.  You must reboot the instance after this recipe runs, which
means a green cluster should ideally be rebooted right after
provisioning.

When we switch to a custom AMI, the second reboot will not be necessary.
Ideally this would be bundled in the default chef recipes.

Includes:

* ixgbevf version 2.16.1 with necessary patches
* Better messaging, use a tmp build directory
* Assets are stored in a bucket that we control
* Ensure the aws-cli is installed