Fix the matterhorn init script to restart until actually successful
Background:

There's a race condition that causes the daemon to not restart correctly
about 10 to 15% of the time. The easiest check is to see if the login
appears after a restart. This modifies the init process to look for
"j_username" in the source of the homepage, which means we're past the
race condition and the daemon is up.

If it can't succeed in 10 attempts, it aborts.

We also relax the matterhorn_availability alerts to require availability
alarms over two periods instead of just one.