Add an instances_started_ok metric and alarm.
By default this metric runs on the monitoring node and is bound to the
AWS/OpsworksCustom namespace and StackId dimension.

It looks at all instances in the stack - if any are in
"connection_lost", "setup_failed", or "start_failed" (which are
non-recoverable error states), report that as an error which will get
thrown to the SNS topic for this cluster.